feat(gatsby-theme-docz): add custom logo support
