+++ core/gatsby-theme-docz/src/components/Logo/index.css
@@ -1,0 +1,1 @@
+@import url('https://fonts.googleapis.com/css?family=Merriweather&display=swap');
