+++ core/gatsby-theme-docz/src/components/Logo/styles.js
@@ -1,0 +1,15 @@
+import './index.css'
+
+export const logo = {
+  fontFamily: 'Merriweather',
+  fontSize: 3,
+}
+
+export const link = {
+  fontWeight: 600,
+  color: 'header.text',
+  textDecoration: 'none',
+  ':hover': {
+    color: 'primary',
+  },
+}
